# library-openapi.yaml
openapi: 3.0.3
info:
  title: L3 Miage PhotoNum
  description: |
    Cette api permet la gestion de la base de donnée d'une société de tirage de photo numérique.
    Elle permet de gérer les clients, les adresses, les commandes, les articles, les impressions, les tirages, les cadres, les calendriers, les albums, les pages, les images et les photos.
  version: 1.0.0
servers:
  - url: "http://localhost:8080"
paths:
...
  /api/v1/Client/{id}/Photo:
    parameters:
      - name: id
        description: id du client
        in: path
        required: true
        schema:
         type: number
         format: int64
    post:
      summary: Creer une nouvelle photo pour un client donné
      operationId: new-client-photo
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/NewPhoto"
      responses:
        201:
          description: Créée
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Photo"
        404:
          description: Le client n'a pas été trouvé
...
  /api/v1/Photo/{id}:
    parameters:
      - name: id
        description: id de la photo
        in: path
        required: true
        schema:
          type: number
          format: int64
    get:
      summary: Trouver une photo
      operationId: find-photo
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Photo"

        404:
          description: La photo n'a pas été trouvée 
    put:
      summary: Mettre à jour photo
      description: La mise à jour sera faite si la photo est trouvée
      operationId: update-photo
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/Photo"
      responses:
        200:
          description: Mis à jour
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Photo"
        400:
          description: Dans le cas où la photo ne peut être validée
        404:
          description: La photo n'a pas été trouvée
    delete:
      summary: Supprimer une photo (et toutes ses adresses)
      operationId: delete-photo
      responses:
        204:
          description: Supprimé
        404:
          description: Le photo n'a pas été trouvée


###############################################################################################################
...
  /api/v1/Article/{id}:
    parameters:
      - name: id
        description: id de l'article
        in: path
        required: true
        schema:
          type: number
          format: int64
    get:
      summary: Trouver un article
      operationId: get-order
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                  $ref: "#/components/schemas/Article"
        404:
          description: L'article n'a pas été trouvé
    put:
      summary: Mettre à jour un article
      operationId: update-order
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/Article"
      responses:
        404:
          description: L'article n'a pas été trouvée
    delete:
      summary: Supprimer un article
      operationId: delete-order
      responses:
        204:
          description: Suppriméee
        404:
          description: L'article n'a pas été trouvée
...
  /api/v1/Commande/{id}/Article:
    parameters:
      - name: id
        description: id de la commande
        in: path
        required: true
        schema:
          type: number
          format: int64
    post:
      summary: Creer un nouveau article pour une commande donnée
      operationId: create-commande-article
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/NewArticle"
      responses:
        201:
          description: Créée
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Article"
        404:
          description: La commande n'a pas été trouvée
        400:
          description: Dans le cas où l'article ne peut être validé


  #Cadre
...
  /Cadre/{id}/Photos:
    parameters:
      - name: id
        description: id du cadre
        in: path
        required: true
        schema:
          type: number
          format: int64
    get:
      summary: Trouver les photos d'un cadre
      operationId: get-cadre-photos
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Photo"
        404:
          description: Le cadre n'a pas été trouvé
        
###############################################################################################################
components:
  schemas:
    BaseClient:
      description: Un client, sans id (pour la creation)
      type: object
      properties:
        nom:
          type: string
        prenoms :
          type : array
          #items : 
        email:
          type : string
        motDePasse:
          type : string
      required:
        - name
        - email
        - motDePasse
      example:
        name: Turc
        prenoms : Julien
        email : julien.turc@etu.univ-grenoble-alpes.fr
        motDepasse : 1234
    Client:
      description: Un client, pour tout usage sauf la création
      allOf:
        - $ref: "#/components/schemas/BaseClient"
        - type: object
          properties:
            id:
              type: integer
              format: int64
            name :
              type: string
            email : 
              type : string
            motDePasse :
              type : string
          required:
            - id
            - name
          example:
            id: 1
            name: Turc
...
    NewPhoto:
      description: Une photo sans id (pour la création)
      type: object
      properties:
        parametresRetouche:
          type: string
        description:
          type: string
        image:
          type: object
      required:
        - description
      example:
        description: cette photo est composée de l'image "cochon.png" et "fleur.png"
    Photo:
      description: Une photo avec id (pour la mise à jour) et son client (pour la réponse)
      allOf:
        - $ref: "#/components/schemas/NewPhoto"
        - type: object
          properties:
            id:
              type: integer
              format: int64
            client:
              type: object
              $ref: "#/components/schemas/Client"
              readOnly: true
          required:
            - id
      example:
        id: 1
        description: cette photo est composée des images "soleil.png" et "fichier.png"

        client:
          - id: 1
            name: Wyss
############################################################################################################
...
    NewArticle:
      description: Un article sans id (pour la création)
      type: object
      properties:
        format:
            type: string
        quantite:
            type: integer
            format: int64
      required:
          - format
          - quantite
      example:
          format: 10X15MAT
          quantite: 3
    Article:
        description: Un article avec id et prix
        type: object
        properties:
          id:
            type: integer
            format: int64
          prix:
            type: integer
            format: int64
          format:
            type: string
          quantite:
            type: integer
            format: int64
        required:
          - id
          - prix
          - format
          - quantite
        example:
          id: 1
          prix: 10.5
          format: 10X15MAT
          quantite: 3

#################################################################################################################
...